validation

+++ app/Http/Controllers/TagController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Tag;
+use Illuminate\Support\Facades\Validator;
+use http\Env\Response;
 use Illuminate\Http\Request;
 
 class TagController extends Controller
@@ -10,13 +12,20 @@ class TagController extends Controller
 
     public function index()
     {
-       $tags = Tag::all();
-       return view('tags.index', ['tags'=>$tags]);
-
+        $tags = Tag::all();
+        return response()->json($tags);
     }
 
     public function store(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+        ]);
+
+        if ($validator->fails()) {
+            return response(['message' => 'error']);
+        }
+
         $tag = new Tag();
         $tag->name = $request->input("name");
         $tag->description = $request->input("description");
@@ -24,12 +33,22 @@ class TagController extends Controller
         return response()->json([
             "message" => "success create"
         ], 201);
-
     }
 
     public function update(Request $request, $id)
     {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+        ]);
+
+        if ($validator->fails()) {
+            return response(['message' => 'error']);
+        }
+
         $tag = Tag::find($id);
+        if (!$tag) return response()->json([
+            'message' => 'tag ' .$id  .   ' not found'
+        ]);
 
         $tag->name = $request->input("name");
         $tag->description = $request->input("description");
@@ -44,7 +63,11 @@ class TagController extends Controller
 
     public function destroy($id)
     {
+
+
         $tag = Tag::find($id);
+
+        if (!$tag) return response()->json(['message' => 'tag  ' .$id  .   ' not found']);
 
         $tag->delete();
 
@@ -52,5 +75,15 @@ class TagController extends Controller
             "message" => "success create"
         ], 200);
 
+    }
+
+
+    public function show($id){
+
+        $tag = Tag::find($id);
+
+        if (!$tag) return response()->json(['message' => 'tag  ' .$id  .   ' not found']);
+
+        return response()->json($tag);
     }
 }
